UPDATE isolate the filter component
