+++ client/src/components/Map/Dot/Dot.css
@@ -1,3 +1,68 @@
-.dot_wrapper{
+.dot {
+  background: var(--color-2);
+  height: 10px;
+  width: 10px;
+  border-radius: 50%;
+  animation-name: pulse;
+  animation-duration: 2s;
+  animation-iteration-count: infinite;
+  animation-timing-function: cubic-bezier(0.075, 0.82, 0.165, 1);
+}
 
+.dot:hover {
+  background: white;
+  transform: scale(5);
+  cursor: default;
+  animation: unset;
+}
+
+.map-img {
+  padding: 20px;
+  color: palevioletred;
+}
+
+.dot:hover .tooltip {
+  visibility: visible;
+  opacity: 1;
+  font-size: 4px;
+  padding: 5px;
+}
+
+.dot:hover .tooltip img {
+  visibility: visible;
+}
+.tooltip img {
+  height: 50px;
+  margin: 0 auto;
+  visibility: hidden;
+  display: block;
+}
+
+.tooltiptext.spec {
+  display: initial;
+  top: 25px;
+}
+
+p.tooltip-money {
+  font-size: 4px;
+  color: white;
+  margin: 0;
+  float: left;
+}
+
+@keyframes pulse {
+  from {
+    -webkit-transform: scale3d(1, 1, 1);
+    transform: scale3d(1, 1, 1);
+  }
+
+  50% {
+    -webkit-transform: scale3d(1.05, 1.05, 1.05);
+    transform: scale3d(2.05, 2.05, 2.05);
+  }
+
+  to {
+    -webkit-transform: scale3d(1, 1, 1);
+    transform: scale3d(1, 1, 1);
+  }
 }
