+++ client/src/components/Map/Dot/Dot.react.jsx
