+++ client/src/components/Map/Map.component.jsx
@@ -3,7 +3,8 @@ import GoogleMapReact from "google-map-react";
 import "./Map.css";
 import * as options from "./map-options";
 import * as projects from "./projects-data.json";
-import Dot from "./Dot/Dot.react";
+import Dot from "./Dot/Dot.react.jsx";
+import Filter from "./Filter/Filter.react.jsx";
 
 export default class Map extends Component {
   state = {
@@ -31,21 +32,10 @@ export default class Map extends Component {
     });
   };
 
-  // mouse hover handler for the spectrum;
-  handleMouseHover = e => {
-    const index = e.target.dataset.index;
-    const types = [
-      "housing",
-      "education",
-      "agriculture",
-      "health",
-      "water",
-      "nutrition",
-      "infancy"
-    ];
-
+  // get the current type from the filter component and set it to the state;
+  setCurrentType = type => {
     this.setState({
-      currentType: types[index]
+      currentType: type
     });
   };
 
@@ -69,73 +59,7 @@ export default class Map extends Component {
         style={{ height: "100vh", width: "100%" }}
         className="map fadeInFast"
       >
-        <div className="filter-spectrum">
-          <ul className="spectrum">
-            <li
-              className="spectrum-item"
-              data-index={0}
-              onMouseEnter={this.handleMouseHover}
-            >
-              <div class="tooltip">
-                <div class="tooltiptext spec">housing projects</div>
-              </div>
-            </li>
-            <li
-              className="spectrum-item"
-              data-index={1}
-              onMouseEnter={this.handleMouseHover}
-            >
-              <div class="tooltip">
-                <div class="tooltiptext spec">education projects</div>
-              </div>
-            </li>
-            <li
-              className="spectrum-item"
-              data-index={2}
-              onMouseEnter={this.handleMouseHover}
-            >
-              <div class="tooltip">
-                <div class="tooltiptext spec">agriculture projects</div>
-              </div>
-            </li>
-            <li
-              className="spectrum-item"
-              data-index={3}
-              onMouseEnter={this.handleMouseHover}
-            >
-              <div class="tooltip">
-                <div class="tooltiptext spec">health projects</div>
-              </div>
-            </li>
-            <li
-              className="spectrum-item"
-              data-index={4}
-              onMouseEnter={this.handleMouseHover}
-            >
-              <div class="tooltip">
-                <div class="tooltiptext spec">water projects</div>
-              </div>
-            </li>
-            <li
-              className="spectrum-item"
-              data-index={5}
-              onMouseEnter={this.handleMouseHover}
-            >
-              <div class="tooltip">
-                <div class="tooltiptext spec">nutrition projects</div>
-              </div>
-            </li>
-            <li
-              className="spectrum-item"
-              data-index={6}
-              onMouseEnter={this.handleMouseHover}
-            >
-              <div class="tooltip">
-                <div class="tooltiptext spec">infancy projects</div>
-              </div>
-            </li>
-          </ul>
-        </div>
+        <Filter setCurrentType={this.setCurrentType} />
         <GoogleMapReact
           options={options}
           bootstrapURLKeys={{ key: "AIzaSyAxYHlwX3Vu7-ygTF2wiB3sjSyFU7mAMJE" }}
