+++ client/src/components/Map/Map.css
@@ -1,164 +1,3 @@
 .map {
   background: var(--color-1);
 }
-
-.dot {
-  background: var(--color-2);
-  height: 10px;
-  width: 10px;
-  border-radius: 50%;
-  animation-name: pulse;
-  animation-duration: 2s;
-  animation-iteration-count: infinite;
-  animation-timing-function: cubic-bezier(0.075, 0.82, 0.165, 1);
-}
-
-.dot:hover {
-  background: white;
-  transform: scale(5);
-  cursor: default;
-  animation: unset;
-}
-
-.map-img {
-  padding: 20px;
-  color: palevioletred;
-}
-
-.spectrum {
-  position: absolute;
-  z-index: 100;
-  margin: 50px 0;
-}
-
-.spectrum:hover li {
-  transition: all 0.5s ease;
-  height: 25px;
-}
-
-.filter-spectrum {
-  display: flex;
-  justify-content: center;
-}
-
-.spectrum-item {
-  background: white;
-  height: 10px;
-  width: 100px;
-  transition: all 0.5s ease;
-  display: inline-block;
-  cursor: pointer;
-}
-
-.spectrum-item:hover .tooltip {
-  visibility: visible;
-  opacity: 1;
-}
-
-.spectrum-item:hover .tooltiptext {
-  font-size: 18px;
-}
-
-.spectrum-item:first-child {
-  border-top-left-radius: 10px;
-  border-bottom-left-radius: 10px;
-  background: linear-gradient(90deg, #e83338, #ff9068);
-}
-.spectrum-item:nth-child(2) {
-  background: linear-gradient(90deg, #ff9068, #ffb75e);
-}
-.spectrum-item:nth-child(3) {
-  background: linear-gradient(90deg, #ffb75e, #8dc26f);
-}
-.spectrum-item:nth-child(4) {
-  background: linear-gradient(90deg, #8dc26f, #64b3f4);
-}
-.spectrum-item:nth-child(5) {
-  background: linear-gradient(90deg, #64b3f4, #6441a5);
-}
-.spectrum-item:nth-child(6) {
-  background: linear-gradient(90deg, #6441a5, #fc67fa);
-}
-.spectrum-item:last-child {
-  border-top-right-radius: 10px;
-  background: linear-gradient(90deg, #fc67fa, #f4c4f3);
-  border-bottom-right-radius: 10px;
-}
-
-.tooltip {
-  text-align: center;
-  width: fit-content;
-  visibility: hidden;
-  background-color: rgba(0, 0, 0, 0.3);
-  width: 6vw;
-  z-index: 1;
-  position: absolute;
-  top: 15px;
-  opacity: 0;
-  transition: opacity 1s;
-  margin-left: -58px;
-}
-
-.tooltip .tooltiptext {
-  font-family: var(--main-font);
-  color: #fff;
-  font-family: var(--main-font);
-  font-size: 6px;
-}
-
-.tooltip .tooltiptext::after {
-  content: "";
-  position: absolute;
-  bottom: 100%;
-  left: 50%;
-  margin-left: -10px;
-  border-width: 10px;
-  border-style: solid;
-  border-color: transparent transparent rgba(0, 0, 0, 0.3) transparent;
-}
-
-.dot:hover .tooltip {
-  visibility: visible;
-  opacity: 1;
-  font-size: 4px;
-  padding: 5px;
-}
-
-.dot:hover .tooltip img {
-  visibility: visible;
-}
-.tooltip img {
-  height: 50px;
-  margin: 0 auto;
-  visibility: hidden;
-  display: block;
-}
-
-.tooltiptext.spec {
-  display: initial;
-  top: 25px;
-}
-
-p.tooltip-money {
-  font-size: 4px;
-  color: white;
-  margin: 0;
-  float: left;
-}
-
-@keyframes pulse {
-  from {
-    -webkit-transform: scale3d(1, 1, 1);
-    transform: scale3d(1, 1, 1);
-  }
-
-  50% {
-    -webkit-transform: scale3d(1.05, 1.05, 1.05);
-    transform: scale3d(2.05, 2.05, 2.05);
-  }
-
-  to {
-    -webkit-transform: scale3d(1, 1, 1);
-    transform: scale3d(1, 1, 1);
-  }
-}
